Rename and move envs.GithubEnv as github.Env
commit-id:57d7d2af
