+++ gitzen/cli.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from gitzen import config, console, envs, git, github
+from gitzen import config, console, git, github
 from gitzen.commands import hook, init, push, status
 from gitzen.console import RealConsoleEnv
 
@@ -13,7 +13,7 @@ def main(args: List[str]) -> None:
     # verify that we are in a git repo or exit
     root_dir = git.root_dir(console_env, git_env)
 
-    github_env: envs.GithubEnv = github.RealGithubEnv()
+    github_env = github.RealGithubEnv()
     cfg = config.load(console_env, root_dir)
 
     for i, arg in enumerate(args):
